build curator client with discoveryProperty
