+++ spring-cloud-huawei-zookeeper/pom.xml
@@ -32,6 +32,5 @@
 
     <modules>
         <module>spring-cloud-huawei-zookeeper-discovery</module>
-        <module>spring-cloud-huawei-zookeeper-common</module>
     </modules>
 </project>
