--- spring-cloud-huawei-zookeeper/spring-cloud-huawei-zookeeper-common/src/main/java/com/huaweicloud/zookeeper/common/ConditionalOnZookeeperEnabled.java
@@ -1,32 +1,0 @@
-/*
-
- * Copyright (C) 2020-2024 Huawei Technologies Co., Ltd. All rights reserved.
-
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.huaweicloud.zookeeper.common;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ ElementType.TYPE, ElementType.METHOD })
-@ConditionalOnProperty(value = "spring.cloud.zookeeper.enabled", matchIfMissing = true)
-public @interface ConditionalOnZookeeperEnabled {
-
-}
