+++ spring-cloud-huawei-zookeeper/spring-cloud-huawei-zookeeper-discovery/pom.xml
@@ -43,10 +43,6 @@
             <artifactId>spring-cloud-huawei-common</artifactId>
         </dependency>
         <dependency>
-            <groupId>com.huaweicloud</groupId>
-            <artifactId>spring-cloud-huawei-zookeeper-common</artifactId>
-        </dependency>
-        <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-actuator</artifactId>
             <scope>provided</scope>
